fix: GCC 15 compatibility for Linux build
- Add -Wno-array-bounds to GCC flag list in 00-Common.cmake
  GCC 15 has a false-positive -Warray-bounds on the LLRefCount /
  std::function pattern used in llpaneloutfitedit.cpp. Without this
  flag, -Werror turns it into a hard error.

- Add static_cast<int> around boost::json::kind in fsdiscordgateway.cpp
  GCC 15 requires explicit enum-to-int cast for printf %d format.
  GCC 11 (upstream dev environment) does not enforce this.

+++ indra/newview/fsdiscordgateway.cpp
@@ -1230,7 +1230,7 @@ void FSDiscordGateway::refreshContactCache()
         }
         else
         {
-            GWLOG("refreshContactCache: friends response is not an array (type=%d)", jv.kind());
+            GWLOG("refreshContactCache: friends response is not an array (type=%d)", static_cast<int>(jv.kind()));
         }
     }
     catch (const std::exception& e)
